Set up Docker and switch from SQLite3 to PostgreSQL
This update sets up a Docker environment for the application, which involves creating a Dockerfile, compose.yaml, and .dockerignore. The database within the Django settings is now configured to use PostgreSQL instead of SQLite3, with the password securely stored and added to environment variables. Newly required dependencies, gunicorn and psycopg2-binary, have also been added to the requirements.txt file.